+++ src/components/plot-layouts/PCPlot/PCPlot.vue
